Added ability to build stubs for JAR files
The createEjbStubs script supports the ability to create EJB stubs
directly with JAR and EAR files. This commit adds two new parameters
(inputFile and updateFile) that allow users to specify an input JAR file
(rather than a list of classes) and an optional destination JAR file.

The user can still configure the plugin to creates stubs for a list of
classes by not specifying inputFile and by providing a list of classes.

Changed version to 1.1.0-SNAPSHOT since this is a new, minor feature.